LOLDataAgent/requirements.txt
langchain
langchain-community
langchain-openai
pymysql
python-dotenv
pyyaml
sqlalchemy
